requirements.txt
Django>=3.2.3,<3.3
djangorestframework>=3.12.4
#psycopg2
#mysqlclient>=2.0
